% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latamMap.R
\docType{data}
\name{latamMap}
\alias{latamMap}
\title{A list of sf objects with Latin American and the Caribbean countries}
\format{
A list of 51 Multipolygon 'sf' objects
}
\source{
\url{https://gadm.org}
}
\usage{
latamMap
}
\description{
Maps used to perform the validation of the original geographical coordinates.
For each country, the map contains the lowest administrative level
available.
}
\keyword{datasets}
