% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\docType{data}
\name{world}
\alias{world}
\title{Simplified world data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 256 rows and 3 columns.
}
\source{
\url{https://biogeo.ucdavis.edu/data/gadm3.6/gadm36_levels_shp.zip}
}
\usage{
world
}
\description{
Simple feature collection with 256 features and 2 fields, corresponding to
GADM countries, downloaded from
"https://biogeo.ucdavis.edu/data/gadm3.6/gadm36_levels_shp.zip", simplified
using sf::st_simplify with a tolerance of 0.01. Country names have been
transformed to lower case and non-ascii characters.
}
\details{
Simple feature collection with 256 features and 2 fields
\describe{
\item{GID_0}{The ISO code of the country}
\item{NAME_0}{The name of the country (lower case, non-ascii removed)}
}
}
\keyword{datasets}
